update files in lesson_2 and lesson_3

+++ module_2/lesson_2/src/solution_2.py
@@ -1,5 +1,5 @@
 number1 = int(input('Введите первое число: '))
 number2 = int(input('Введите второе число: '))
 number3 = int(input('Введите третье число: '))
-average: int = (number1 + number2 + number3) / 3
+average: float = (number1 + number2 + number3) / 3
 print(average)
